moved user related types to types/user
