+++ src/pages/two-factor/setup/index.tsx
@@ -4,11 +4,7 @@ import { CodeBlock, FreeFlowText } from 'components/RecoveryKeyModal';
 import { DeadCenter } from 'pages/gallery';
 import React, { useContext, useEffect, useState } from 'react';
 import { Button, Card } from 'react-bootstrap';
-import {
-    enableTwoFactor,
-    setupTwoFactor,
-    TwoFactorSecret,
-} from 'services/userService';
+import { enableTwoFactor, setupTwoFactor } from 'services/userService';
 import styled from 'styled-components';
 import constants from 'utils/strings/constants';
 import Container from 'components/Container';
@@ -19,6 +15,7 @@ import { encryptWithRecoveryKey } from 'utils/crypto';
 import { setData, LS_KEYS, getData } from 'utils/storage/localStorage';
 import { AppContext, FLASH_MESSAGE_TYPE } from 'pages/_app';
 import { PAGES } from 'types';
+import { TwoFactorSecret } from 'types/user';
 
 enum SetupMode {
     QR_CODE,
