+++ src/pages/verify/index.tsx
@@ -12,8 +12,6 @@ import {
     getOtt,
     logoutUser,
     clearFiles,
-    EmailVerificationResponse,
-    User,
     putAttributes,
 } from 'services/userService';
 import { setIsFirstLogin } from 'utils/storage';
@@ -21,7 +19,9 @@ import SubmitButton from 'components/SubmitButton';
 import { clearKeys } from 'utils/storage/sessionStorage';
 import { AppContext } from 'pages/_app';
 import LogoImg from 'components/LogoImg';
-import { KeyAttributes, PAGES } from 'types';
+import { PAGES } from 'types';
+import { User } from '@sentry/nextjs';
+import { KeyAttributes, EmailVerificationResponse } from 'types/user';
 
 interface formValues {
     ott: string;
