+++ src/utils/crypto/index.ts
@@ -1,5 +1,5 @@
 import { KEK } from 'pages/generate';
-import { KeyAttributes } from 'types';
+import { KeyAttributes } from 'types/user';
 import * as Comlink from 'comlink';
 import { runningInBrowser } from 'utils/common';
 import { SESSION_KEYS, setKey } from 'utils/storage/sessionStorage';
